requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
pycodestyle==2.9.1
sqlparse==0.4.2
toml==0.10.2
